+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Potential problems with garbage collection of the `result` array and other `VALUE`s
 
+### Added
+
+- Report `bytes_consumed` in `ParseError`s
+
 ## [0.2.0] - 2024-12-27
 
 ### Fixed